Rewrite of CSVWriter and CSVReader. Both, reader and writer, collected all data in memory, leading to high memory usage, before writing the data to a file or passing the read data on for processing. Changed the two classes such that CSV-rows are read line by line, only having the current line in memory, as well as write CSV-row after row instead of first caching all in memory. This required major changes in the post processing classes as those used the CSV-Readers and -Writers most.
I've also fixed a but in the class PTObjective: the row "departureId" was sometimes accessed as "departure". Not sure, what effect this bug had...

The old CSV-Writer added a ';' after each written column, even after the last. I've removed that last ';' -- according to the CSV standard, this should not be necessary. I hope VISUM thinks the same...